renamed to SliderBlock
