+++ src/components/SliderBlock/SliderBlock.vue
